Modificación en los Models

+++ src/main/java/com/library/project/web/repository/IAutorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.library.project.web.models.Autor;
 
-public interface IAutorRepository extends JpaRepository<Autor, Long> {
+public interface IAutorRepository extends JpaRepository<Autor, Integer> {
 
 }
