+++ src/main/java/com/library/project/web/repository/ICarreraRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.library.project.web.models.Carrera;
 
-public interface ICarreraRepository extends JpaRepository<Carrera, Long> {
+public interface ICarreraRepository extends JpaRepository<Carrera, Integer> {
 
 }
