+++ src/main/java/com/library/project/web/repository/IEstudianteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.library.project.web.models.Estudiante;
 
-public interface IEstudianteRepository  extends JpaRepository<Estudiante, Long>{
+public interface IEstudianteRepository  extends JpaRepository<Estudiante, Integer>{
 
 }
